refactor: refactor contracts folder structure
