+++ src/common/memory-repository.ts
@@ -1,4 +1,4 @@
-import type { IRepository, IConditionCallback } from '../contract/i-repository'
+import type { IRepository, IConditionCallback } from '../contract/repository.contract'
 
 function byStrictEqual<TItem>(target: TItem): IConditionCallback<TItem> {
   return (item: TItem): boolean => item === target
@@ -9,7 +9,7 @@ function byStrictEqual<TItem>(target: TItem): IConditionCallback<TItem> {
  */
 class MemoryRepository<TItem> implements IRepository<TItem> {
   _collection: TItem[] = []
-  
+
   async add(...items: TItem[]) {
     this._collection.push(...items)
   }
@@ -17,25 +17,25 @@ class MemoryRepository<TItem> implements IRepository<TItem> {
   async get(condition: IConditionCallback<TItem>) {
     return this._collection.filter(condition)
   }
-  
+
   async getOnce(condition: IConditionCallback<TItem>) {
     const entity = this._collection.find(condition)
     return entity || null
   }
-  
+
   async getAll() {
     return [...this._collection]
   }
-  
+
   async remove(condition: IConditionCallback<TItem>) {
     const toRemoveList = await this.get(condition)
-    
+
     toRemoveList.forEach((toRemove) => {
       const index = this._collection.findIndex(byStrictEqual(toRemove))
       this._collection.splice(index, 1)
     })
   }
-  
+
   static create<T>(): MemoryRepository<T> {
     return new MemoryRepository()
   }
