+++ src/contract/model.contract.ts
@@ -1,0 +1,4 @@
+import { IResponsable } from "./responsable.contract";
+import { IStorable } from "./storable.contract";
+
+export interface IModel extends IStorable, IResponsable {}
