+++ src/resources/columns/column.model.ts
@@ -1,6 +1,6 @@
 import { v4 as uuid } from 'uuid'
-import { IResponsable } from '../../contract/i-responsable'
-import { IStorable } from '../../contract/i-storable'
+import { IResponsable } from '../../contract/responsable.contract'
+import { IStorable } from '../../contract/storable.contract'
 
 export interface ColumnLike {
   id?: string,
@@ -24,28 +24,28 @@ export class Column implements ColumnLike, IStorable, IResponsable {
     this.title = title
     this.order = order
   }
-  
+
   toStorage() {
     return Column.toStorage(this)
   }
-  
+
   toResponse() {
     return Column.toResponse(this)
   }
-  
+
   static create(columnLike: ColumnLike) {
     return new Column(columnLike)
   }
-  
+
   static getId(column: Column) {
     return column.id
   }
-  
+
   static toStorage(column: Column) {
     const { id, title, order } = column
     return { id, title, order }
   }
-  
+
   static toResponse(column: Column) {
     const { id, title, order } = column
     return { id, title, order }
